:truck: rename: MyPageTalk ⇒ MyPageTalkList 이름 변경 및 파일 경로 변경 #306
